sore/sore/asdfghjkl;ASDFGHJKL/requirements.txt
# Core dependencies
python-dotenv>=1.0.0
structlog>=24.0.0
aiokafka>=0.8.0
httpx[http2]>=0.24.0
tenacity>=8.2.0

# Configuration and Validation
pydantic>=2.0,<3.0
pydantic-settings>=2.0
PyYAML>=6.0.0

# Data Processing
feedparser>=6.0.0
spacy>=3.0.0

# Messaging
redis>=4.7.0

# Testing
pytest>=8.0
pytest-asyncio>=0.21.0
respx>=0.23.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Development
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0
types-redis>=4.10.0

